mod com att

+++ desafios/d008.py
@@ -1,5 +1,5 @@
 aluno = str(input('Nome do aluno: '))
-n1 = int(input('1º nota: '))
-n2 = int(input('2º nota: '))
+n1 = float(input('1º nota: '))
+n2 = float(input('2º nota: '))
 media = (n1+n2)/2
 print('a media de {} foi de {:.2f}'.format(aluno, media))
